pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orderbook-alpha-signals"
version = "1.0.0"
description = "Quantitative trading research toolkit for orderbook signal analysis"
authors = [{name = "Mathieu Delehaye", email = "mathieu.delehaye@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["quantitative-finance", "trading", "orderbook", "signals", "backtesting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "matplotlib>=3.5.0", 
    "seaborn>=0.11.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.60.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
    "jupyter>=1.0.0",
    "nbval>=0.9.0"
]

performance = [
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.60.0"
]

security = [
    "bandit>=1.7.0", 
    "safety>=2.0.0"
]

docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "pdoc3>=0.10.0"
]

[project.urls]
Homepage = "https://github.com/mathieudelehaye/orderbook-alpha-signals"
Repository = "https://github.com/mathieudelehaye/orderbook-alpha-signals"
Issues = "https://github.com/mathieudelehaye/orderbook-alpha-signals/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]